requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==25.4.0
babel==2.17.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
daff==1.4.2
databricks-connect==17.2.*
databricks-sdk==0.63.0
databricks-sql-connector==4.0.5
dbt-adapters==1.17.3
dbt-common==1.33.0
dbt-core==1.10.13
dbt-databricks==1.10.14
dbt-extractor==0.6.0
dbt-protos==1.0.382
dbt-semantic-interfaces==0.9.0
dbt-spark==1.9.3
deepdiff==8.6.1
et-xmlfile==2.0.0
google-auth==2.42.1
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
importlib-metadata==8.7.0
iniconfig==2.3.0
isodate==0.6.1
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.3.0
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.6.0
leather==0.4.0
lz4==4.4.4
markupsafe==3.0.3
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.2
networkx==3.5
numpy==2.3.4
oauthlib==3.3.1
openpyxl==3.1.5
orderly-set==5.5.0
packaging==25.0
pandas==2.2.3
parsedatetime==2.6
pathspec==0.12.1
pillow==12.0.0
pluggy==1.6.0
protobuf==6.33.0
py4j==0.10.9.9
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.3
pydantic-core==2.41.4
pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pyyaml==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
snowplow-tracker==1.1.0
sqlparams==6.2.0
sqlparse==0.5.3
svgwrite==1.4.3
text-unidecode==1.3
thrift==0.20.0
tree==0.2.4
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
zstandard==0.25.0
